Many-To-Many

+++ src/main/java/com/projects/countrycode/repodao/CountryRepository.java
@@ -2,6 +2,7 @@ package com.projects.countrycode.repodao;
 
 import com.projects.countrycode.domain.Country;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
@@ -9,9 +10,8 @@ import java.util.List;
 //DAO: Предоставляет интерфейс для доступа к данным в базе данных.
 
 @Repository
-public interface CountryRepository extends JpaRepository<Country, Long> { //Id as primary key
+public interface CountryRepository extends JpaRepository<Country, Integer> { //Id as primary key
     Country findByCountryName(String countryName);
     List<Country> findByPhoneCode(Long phoneCode);
-
 }
